perf(tests): the sample projects publish once per family, not once per sample
The 14 sample-project ITs were the same journey 14 times - clone a
dirigiblelabs/* repo through the Git perspective, publishAll, verify - each
paying its own Dirigible boot, Chrome session and publish cycle (1088s on the
h2 leg, the biggest remaining block in the samples shard).

SampleProjectsIT now clones a FAMILY of repos into one workspace, publishes
once, and lets each subclass verify one sample per @Test under AFTER_CLASS:
TypeScriptSampleProjectsIT (8 samples) and JavaSampleProjectsIT (5 samples)
replace 13 classes, so 13 boots become 2. SampleLibraryLocalNativeAppIT moves
onto the new base but keeps its own instance - it spawns a real OS process.
Every assertion is carried over unchanged.

Two families rather than one class because sample-entity-decorators and
sample-java-entity-decorators both own the SAMPLE_COUNTRY table and both
CSVIM-seed it; the split keeps them in separate instances and needs no change
to any sample repo. The preflight for the Java family came back clean - all 17
client classes sit in distinct demo.* packages with distinct simple names, so
the shared javac batch and bean container have nothing to collide on.

The one-time publish is a guarded @BeforeEach, not a @BeforeAll on a
@TestInstance(PER_CLASS) class: PER_CLASS autowires the instance - booting the
platform - before @BeforeAll runs, so IntegrationTest's cleaner then deleted
the Dirigible folder under the running instance and the next sync pass reaped
the platform's own registry (Definition deleted:
/shell-ide/extensions/shell.extension), leaving the IDE with no perspectives.
The guard is keyed by class, not a boolean: the field is shared by every
subclass, and a flag would let the second family skip its own publish.

Part of #6710.
