+++ components/engine/engine-native-apps/CLAUDE.md
@@ -37,4 +37,4 @@ User projects can declare a `*.nativeapp` JSON file that turns an external web s
 - **Gotchas worth remembering.**
   - `Python 3.14` on macOS has a bind regression in `http.server.HTTPServer` — don't use Python for embedded test servers; use `java Server.java`.
   - `ServerSocket(port, 0, InetAddress.getLoopbackAddress())` can succeed on loopback even when another listener is bound to `0.0.0.0:port` on macOS — this is why `PortResolver.isBindable` deliberately probes the wildcard interface (`new ServerSocket(port, 0, null)`). Keep it that way; reverting to a loopback probe will reintroduce silent `EADDRINUSE` failures when the parent platform (or any other process) is already on the preferred port.
-  - `Spring Boot's DevTools` (or whatever `FileSystemWatcher` wires up) can deadlock against `sun.nio.fs.PollingWatchService$PollingWatchKey` on macOS during full IT runs (last seen on `JavaLspIT`, `CreateNewFileIT`, `JavaEntityDecoratorsSampleProjectIT`). This is NOT a native-apps regression. When the full reactor IT run hangs, jstack the surefire JVM and check for `Found one Java-level deadlock` between `main` and `FileSystemWatcher`.
+  - `Spring Boot's DevTools` (or whatever `FileSystemWatcher` wires up) can deadlock against `sun.nio.fs.PollingWatchService$PollingWatchKey` on macOS during full IT runs (last seen on `JavaLspIT`, `CreateNewFileIT`, `JavaSampleProjectsIT`). This is NOT a native-apps regression. When the full reactor IT run hangs, jstack the surefire JVM and check for `Found one Java-level deadlock` between `main` and `FileSystemWatcher`.
