+++ tests/tests-framework/src/main/java/org/eclipse/dirigible/tests/framework/util/SynchronizationUtil.java
@@ -47,7 +47,7 @@ public class SynchronizationUtil {
      * Wait until the synchronizer stays idle for a full quiet period. A plain idle check is not enough
      * right after a publish: the registry file watcher (a polling watcher on some platforms) delivers
      * trailing change events SECONDS after the copy, scheduling one more cycle - which may re-register
-     * data-store entities and rebuild their tables mid-test (StoreAPISampleProjectIT saw its
+     * data-store entities and rebuild their tables mid-test (the store-api sample verification saw its
      * just-inserted rows vanish exactly this way). Requiring the idle condition to HOLD for ten seconds
      * absorbs those late events before the test proceeds.
      */
