--- tests/tests-integrations/src/main/java/org/eclipse/dirigible/integration/tests/ui/tests/sample/JavaExtensionDecoratorSampleProjectIT.java
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2010-2026 Eclipse Dirigible contributors
- *
- * All rights reserved. This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v2.0 which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v20.html
- *
- * SPDX-FileCopyrightText: Eclipse Dirigible contributors SPDX-License-Identifier: EPL-2.0
- */
-package org.eclipse.dirigible.integration.tests.ui.tests.sample;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.containsString;
-
-public class JavaExtensionDecoratorSampleProjectIT extends SampleProjectRepositoryIT {
-
-    private static final String PROJECT = "sample-java-extension-decorator";
-    private static final String EXTENSION_CONSUMER_BASE = "/services/java/" + PROJECT + "/demo/extension/ExtensionConsumer";
-    private static final String INJECTING_CONSUMER_BASE = "/services/java/" + PROJECT + "/demo/extension/InjectingConsumer";
-
-    @Override
-    protected String getRepositoryURL() {
-        return "https://github.com/dirigiblelabs/sample-java-extension-decorator.git";
-    }
-
-    @Override
-    protected void verifyProject() {
-        restAssuredExecutor.execute(() -> {
-            // Style 1 — Extensions.find(SampleExtensionPoint.class) returns the SampleContribution
-            // instances; the consumer maps each via describe(), so the body carries the contribution's
-            // own string. Asserting on it also verifies the cast — only a real implementor reaches it.
-            given().when()
-                   .get(EXTENSION_CONSUMER_BASE + "/contributions")
-                   .then()
-                   .statusCode(200)
-                   .body(containsString("Hello from SampleContribution!"));
-
-            // Style 2 — collection injection: the container populates the controller's
-            // List<SampleExtensionPoint> with every @Component contribution, no explicit lookup.
-            given().when()
-                   .get(INJECTING_CONSUMER_BASE + "/injected-contributions")
-                   .then()
-                   .statusCode(200)
-                   .body(containsString("Hello from SampleContribution!"));
-        });
-    }
-
-}
